pyproject.toml
[tool.poetry]
name = "aasc_sentence_search_api"
version = "0.1.0"
description = ""
authors = ["Kaito Sugimoto <hellorusk1998@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.66.0"
transformers = "^4.8.2"
flake8 = "^3.9.2"
torch = "^1.9.0"
faiss-cpu = "^1.7.1"
uvicorn = {extras = ["standard"], version = "^0.14.0"}
pandas = "^1.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
